fix(core): correct LD2450 protocol bugs and rewrite decoder tests
Three spec violations inherited from the Gemini-generated spatialtracker
source, all confirmed against HLK-LD2450 V1.03 spec (Table 10, page 12):

1. Sign convention reversed — MSB=1 is positive (raw − 0x8000), not negative
2. Target record is 8 bytes, not 10 — removed two phantom status bytes
3. Speed field is cm/s, not mm/s — divide by 100, not 1000

StaticClutterCalibrator: fix Int→Double type mismatch on avgResolution init.

Test rewrite: correct encodeSignedMag, 30-byte frame layout (AA FF 03 00
header + 3×8-byte targets + 55 CC tail), cm/s speed units, golden frame
from spec page 12 (x=−782 mm, y=+1713 mm, v=−0.16 m/s, res=320 mm),
and explicit config-frame-rejection test.

+++ android/core/src/main/kotlin/com/wscanplus/core/sensor/StaticClutterCalibrator.kt
@@ -110,7 +110,7 @@ class StaticClutterCalibrator {
                     centerX = target.xMm,
                     centerY = target.yMm,
                     hits = 0,
-                    avgResolution = target.resolutionMm,
+                    avgResolution = target.resolutionMm.toDouble(),
                 )
             }
         cell.hits++
